Refactor example output config
Put all the config parsing into shared.h so it is shared among the examples.
