+++ examples/rotation.c
@@ -22,7 +22,7 @@
 #include "cat.h"
 
 struct sample_state {
-	struct wl_list config;
+	struct example_config *config;
 	struct wlr_renderer *renderer;
 	struct wlr_texture *cat_texture;
 };
@@ -30,12 +30,6 @@ struct sample_state {
 struct output_data {
 	float x_offs, y_offs;
 	float x_vel, y_vel;
-};
-
-struct output_config {
-	char *name;
-	enum wl_output_transform transform;
-	struct wl_list link;
 };
 
 static void handle_output_frame(struct output_state *output, struct timespec *ts) {
@@ -81,7 +75,7 @@ static void handle_output_add(struct output_state *output) {
 	struct sample_state *state = output->compositor->data;
 
 	struct output_config *conf;
-	wl_list_for_each(conf, &state->config, link) {
+	wl_list_for_each(conf, &state->config->outputs, link) {
 		if (strcmp(conf->name, output->output->name) == 0) {
 			wlr_output_transform(output->output, conf->transform);
 			break;
@@ -126,76 +120,9 @@ static void handle_keyboard_key(struct keyboard_state *kbstate,
 	}
 }
 
-static void usage(const char *name, int ret) {
-	fprintf(stderr,
-		"usage: %s [-d <name> [-r <rotation> | -f]]*\n"
-		"\n"
-		" -o <output>    The name of the DRM display. e.g. DVI-I-1.\n"
-		" -r <rotation>  The rotation counter clockwise. Valid values are 90, 180, 270.\n"
-		" -f             Flip the output along the vertical axis.\n", name);
-
-	exit(ret);
-}
-
-static void parse_args(int argc, char *argv[], struct wl_list *config) {
-	struct output_config *oc = NULL;
-
-	int c;
-	while ((c = getopt(argc, argv, "o:r:fh")) != -1) {
-		switch (c) {
-		case 'o':
-			oc = calloc(1, sizeof(*oc));
-			oc->name = optarg;
-			oc->transform = WL_OUTPUT_TRANSFORM_NORMAL;
-			wl_list_insert(config, &oc->link);
-			break;
-		case 'r':
-			if (!oc) {
-				fprintf(stderr, "You must specify an output first\n");
-				usage(argv[0], 1);
-			}
-
-			if (oc->transform != WL_OUTPUT_TRANSFORM_NORMAL
-					&& oc->transform != WL_OUTPUT_TRANSFORM_FLIPPED) {
-				fprintf(stderr, "Rotation for %s already specified\n", oc->name);
-				usage(argv[0], 1);
-			}
-
-			if (strcmp(optarg, "90") == 0) {
-				oc->transform += WL_OUTPUT_TRANSFORM_90;
-			} else if (strcmp(optarg, "180") == 0) {
-				oc->transform += WL_OUTPUT_TRANSFORM_180;
-			} else if (strcmp(optarg, "270") == 0) {
-				oc->transform += WL_OUTPUT_TRANSFORM_270;
-			} else {
-				fprintf(stderr, "Invalid rotation '%s'\n", optarg);
-				usage(argv[0], 1);
-			}
-			break;
-		case 'f':
-			if (!oc) {
-				fprintf(stderr, "You must specify an output first\n");
-				usage(argv[0], 1);
-			}
-
-			if (oc->transform >= WL_OUTPUT_TRANSFORM_FLIPPED) {
-				fprintf(stderr, "Flip for %s already specified\n", oc->name);
-				usage(argv[0], 1);
-			}
-
-			oc->transform += WL_OUTPUT_TRANSFORM_FLIPPED;
-			break;
-		case 'h':
-		case '?':
-			usage(argv[0], c != 'h');
-		}
-	}
-}
-
 int main(int argc, char *argv[]) {
 	struct sample_state state = {0};
-	wl_list_init(&state.config);
-	parse_args(argc, argv, &state.config);
+	state.config = parse_args(argc, argv);
 
 	struct compositor_state compositor = { 0 };
 	compositor.data = &state;
@@ -223,8 +150,5 @@ int main(int argc, char *argv[]) {
 	wlr_texture_destroy(state.cat_texture);
 	wlr_renderer_destroy(state.renderer);
 
-	struct output_config *ptr, *tmp;
-	wl_list_for_each_safe(ptr, tmp, &state.config, link) {
-		free(ptr);
-	}
+	example_config_destroy(state.config);
 }
